libmetrics: cast test types for unsigned types
avoid type mismatchs where unsigned is not 32bit

+++ libmetrics/tests/test-metrics.c
@@ -137,7 +137,7 @@ main (void)
 		  fprintf (stderr, "%d (g_int32)", (int) val.int32);
 		  break;
 		case g_uint32:
-		  fprintf (stderr, "%u (g_uint32)", val.uint32);
+		  fprintf (stderr, "%u (g_uint32)", (unsigned int)val.uint32);
 		  break;
 		case g_float:
 		  fprintf (stderr, "%f (g_float)", val.f);
@@ -146,7 +146,7 @@ main (void)
 		  fprintf (stderr, "%f (g_double)", val.d);
 		  break;
 		case g_timestamp:
-		  fprintf (stderr, "%u (g_timestamp)", val.uint32);
+		  fprintf (stderr, "%u (g_timestamp)", (unsigned)val.uint32);
 		  break;
 		}
 	    }
